[ct] Move mock trajectories to test utils
