+++ verl/utils/test_utils/mock_trajectories.py
